+++ src/entities/payment/index.ts
@@ -1,6 +1,8 @@
+import type { GetPaymentsByUserQuery } from '@/shared/api/graphql/__generated__/graphql'
+
+export { GetPaymentsByUserDocument } from './api/documents'
 export { usePaymentsByUserQuery } from './api/usePaymentsByUserQuery'
-export type {
-  GetPaymentsByUserInput,
-  PaymentPaginationModel,
-  SubscriptionByPaymentModel,
-} from '@/shared/api/graphql/client'
+export type { GetPaymentsByUserQueryVariables as GetPaymentsByUserInput } from '@/shared/api/graphql/__generated__/graphql'
+
+export type PaymentPaginationModel = GetPaymentsByUserQuery['getPaymentsByUser']
+export type SubscriptionByPaymentModel = PaymentPaginationModel['items'][number]