+++ src/features/users-list/api/documents.ts
@@ -1,0 +1,7 @@
+import { graphql } from '@/shared/api/graphql/__generated__/gql'
+
+export const RemoveUserDocument = graphql(`
+  mutation RemoveUser($userId: Int!) {
+    removeUser(userId: $userId)
+  }
+`)